ADCS/FSW: Stop using doubles or FRAM

+++ src/devmods/adcsfsw_test_14/rtwtypes.h
@@ -57,14 +57,14 @@ typedef unsigned int uint16_T;
 typedef long int32_T;
 typedef unsigned long uint32_T;
 typedef float real32_T;
-typedef double real64_T;
+typedef float real64_T;
 
 /*===========================================================================*
  * Generic type definitions: boolean_T, char_T, byte_T, int_T, uint_T,       *
  *                           real_T, time_T, ulong_T.                        *
  *===========================================================================*/
-typedef double real_T;
-typedef double time_T;
+typedef float real_T;
+typedef float time_T;
 typedef unsigned char boolean_T;
 typedef int int_T;
 typedef unsigned int uint_T;
